Zustand (#1)
* ad zustand base

* adding store modifiers

* adding core funcitonality

* add storage getters

* addign action update

* sf

* add fixes to zustand

* add sizes to the store

* add toolbar stolre

* remove api key

* add toast message

* toolbar optimization

* prevent rerenders for lines

* updating equality fn

* add aditional fixes

* prevent more rerenders

* variable change

* atomic configs

* refactor colorMap and backColorMap

* drag and drop more performance

* fixing logo

* refactor appjs

* actions fix

* refactor main store

* getting rid of useeffect

* sf

* less re renders main

* join removes

* avoid re renders label

* optimizing backgrounds

* folder restructure

---------

+++ src/CoverBoard/Toolbar/ToolbarShare/ToolbarSharePopover.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useMemo, useEffect } from 'react';
+import React, { useState, useMemo } from 'react';
 import {
   Grid,
   TextareaAutosize,
@@ -12,7 +12,14 @@ import { Close as CloseIcon } from '@mui/icons-material';
 
 import { LocalStorageData, ToolConfigIDs, DEFAULT_KEY } from 'types';
 import { NavigateFunction } from 'react-router-dom';
-import { addPrefix, clearHash, haxPrefix, removePrefix, setHash } from 'utils';
+import {
+  addPrefix,
+  clearHash,
+  DEFAULT_STORAGE,
+  haxPrefix,
+  removePrefix,
+  setHash,
+} from 'utils';
 import { CommonDialog } from 'components';
 
 interface SaveProps {
@@ -41,10 +48,6 @@ export const ToolbarSharePopover: React.FC<SaveProps> = ({
   const [newSave, setNewSave] = useState('');
 
   setHash(ToolConfigIDs.SHARE);
-
-  useEffect(() => {
-    setJsonData(JSON.stringify(instance, null, 4));
-  }, [instance]);
 
   const handleCopyText = () => {
     try {
@@ -98,6 +101,7 @@ export const ToolbarSharePopover: React.FC<SaveProps> = ({
 
     if (value) {
       navigate(`/${newSave}`);
+      // navigate
       setNewSave('');
       handleClose();
     }
@@ -123,13 +127,22 @@ export const ToolbarSharePopover: React.FC<SaveProps> = ({
                   color={saveId === currentSave ? 'primary' : 'default'}
                   onClick={() => {
                     navigate(`/${currentSave}#${ToolConfigIDs.SHARE}`);
+                    const data = window.localStorage.getItem(
+                      addPrefix(currentSave),
+                    );
+                    if (data) {
+                      setJsonData(JSON.stringify(JSON.parse(data), null, 4));
+                    }
                   }}
                   onDelete={
                     showDelete
                       ? (evt) => {
                           evt.preventDefault();
 
-                          if (saveId === DEFAULT_KEY) {
+                          if (
+                            saveId === DEFAULT_KEY &&
+                            currentSave === saveId
+                          ) {
                             handleDeleteElements();
                             handleClose();
                             return;
@@ -147,6 +160,18 @@ export const ToolbarSharePopover: React.FC<SaveProps> = ({
 
                           if (saveId === currentSave) {
                             navigate(`/${DEFAULT_KEY}#${ToolConfigIDs.SHARE}`);
+
+                            const defaultData =
+                              window.localStorage.getItem(DEFAULT_STORAGE);
+                            if (defaultData) {
+                              setJsonData(
+                                JSON.stringify(
+                                  JSON.parse(defaultData),
+                                  null,
+                                  4,
+                                ),
+                              );
+                            }
                           }
                         }
                       : undefined
